+++ playground/run_v4.py
@@ -28,6 +28,7 @@ def audio_postprocess(
         max_audio = np.abs(audio).max()  # 简单防止16bit爆音
         if max_audio > 1:
             audio /= max_audio
+        audio = audio.astype(np.float32)
         audio = np.concatenate([audio, zero_wav], axis=0)
         audios[i] = audio
 
@@ -60,7 +61,7 @@ def audio_preprocess(audio_path):
     """Get HuBERT features for the audio file"""
     waveform = load_audio(audio_path)
     ort_session = ort.InferenceSession(MODEL_PATH + "_export_audio_preprocess.onnx")
-    ort_inputs = {ort_session.get_inputs()[0].name: waveform.numpy()}
+    ort_inputs = {ort_session.get_inputs()[0].name: waveform.numpy().astype(np.float16)}
     [hubert_feature, spectrum, sv_emb, mel2_v4] = ort_session.run(None, ort_inputs)
     return hubert_feature, spectrum, sv_emb, mel2_v4
 
@@ -68,7 +69,7 @@ def preprocess_text(text:str):
     preprocessor = TextPreprocessorOnnx(ROBERTA_PATH)
     [phones, bert_features, norm_text] = preprocessor.segment_and_extract_feature_for_text(text, 'all_zh', 'v2')
     phones = np.expand_dims(np.array(phones, dtype=np.int64), axis=0)
-    return phones, bert_features.T.astype(np.float32)
+    return phones, bert_features.T.astype(np.float16)
 
 
 # input_phones_saved = np.load("playground/ref/input_phones.npy")
@@ -86,9 +87,9 @@ def preprocess_text(text:str):
 # audio_prompt_hubert_saved = np.load("playground/ref/audio_prompt_hubert.npy").astype(np.float32)
 
 top_k = np.array([TOP_K], dtype=np.int64)
-top_p = np.array([TOP_P], dtype=np.float32)
-repetition_penalty = np.array([REPETITION_PENALTY], dtype=np.float32)
-temperature = np.array([TEMPERATURE], dtype=np.float32)
+top_p = np.array([TOP_P], dtype=np.float16)
+repetition_penalty = np.array([REPETITION_PENALTY], dtype=np.float16)
+temperature = np.array([TEMPERATURE], dtype=np.float16)
 
 t2s_init_stage = ort.InferenceSession(MODEL_PATH+"_export_t2s_init_stage.onnx")
 # t2s_init_step = ort.InferenceSession(MODEL_PATH+"_export_t2s_init_step.onnx")
@@ -100,7 +101,7 @@ t2s_init_stage = ort.InferenceSession(MODEL_PATH+"_export_t2s_init_stage.onnx")
     "ref_text_bert": ref_bert,
     "hubert_ssl_content": audio_prompt_hubert,
 })
-empty_tensor = np.empty((1,0,512)).astype(np.float32)
+empty_tensor = np.empty((1,0,512)).astype(np.float16)
 
 t2s_stage_decoder = ort.InferenceSession(MODEL_PATH+"_export_t2s_stage_decoder.onnx")
 y, k, v, y_emb, logits, samples = t2s_stage_decoder.run(None, {
@@ -146,7 +147,7 @@ pred_semantic = np.expand_dims(y[:, -idx:], axis=0)
 
 vits = ort.InferenceSession(MODEL_PATH+"_export_vits.onnx")
 
-speed = np.array([SPEED], dtype=np.float32)
+speed = np.array([SPEED], dtype=np.float16)
 fea_ref, fea_todo, chunk_len, mel2_v4 = vits.run(None, {
     'hubert_ssl_content': audio_prompt_hubert,
     'spectrum': spectrum,
@@ -162,14 +163,14 @@ def cfm_onnx(fea, mel, sample_steps):
     # Implement the ONNX export logic for CFM here
     cfm_onnx = ort.InferenceSession('onnx/v4_export/v4_export_cfm.onnx')
     mu_length = fea.shape[1]
-    x_last = np.empty((0, 100, mu_length), dtype=np.float32)
-    temperature = np.array([1.0], dtype=np.float32)
+    x_last = np.empty((0, 100, mu_length), dtype=np.float16)
+    temperature = np.array([1.0], dtype=np.float16)
     n_timesteps = np.array([sample_steps], dtype=np.int64)
     i_timestep = np.array([0], dtype=np.int64)
     mu = fea
     prompt = mel
-    text_cache = np.empty((0, mu_length, 512), dtype=np.float32)
-    dt_cache = np.empty((0, 1024), dtype=np.float32)
+    text_cache = np.empty((0, mu_length, 512), dtype=np.float16)
+    dt_cache = np.empty((0, 1024), dtype=np.float16)
     for _ in tqdm(range(sample_steps)):
         x_last, text_cache, dt_cache = cfm_onnx.run(None, {
             'x_last': x_last,
